+++ modulus/models/graphcast/graph_cast_processor.py
@@ -16,7 +16,9 @@
 
 from typing import Union
 
+import torch
 import torch.nn as nn
+import transformer_engine as te
 from dgl import DGLGraph
 from torch import Tensor
 
@@ -177,3 +179,61 @@ class GraphCastProcessor(nn.Module):
             )
 
         return efeat, nfeat
+
+
+class GraphCastProcessorGraphTransformer(nn.Module):
+    """Processor block used in GenCast operating on a latent space
+    represented by hierarchy of icosahedral meshes.
+
+    Parameters
+    ----------
+    attn_mask : torch.Tensor
+        Attention mask to be applied within the transformer layers.
+    processor_layers : int, optional (default=16)
+        Number of processing layers.
+    input_dim_nodes : int, optional (default=512)
+        Dimension of the input features for each node.
+    hidden_dim : int, optional (default=512)
+        Dimension of the hidden features within the transformer layers.
+    """
+
+    def __init__(
+        self,
+        attention_mask: torch.Tensor,
+        num_attention_heads: int = 4,
+        processor_layers: int = 16,
+        input_dim_nodes: int = 512,
+        hidden_dim: int = 512,
+    ):
+        super().__init__()
+        self.num_attention_heads = num_attention_heads
+        self.hidden_dim = hidden_dim
+        self.attention_mask = torch.tensor(attention_mask, dtype=torch.bool)
+        self.register_buffer("mask", self.attention_mask, persistent=False)
+
+        layers = [
+            te.pytorch.TransformerLayer(
+                hidden_size=input_dim_nodes,
+                ffn_hidden_size=hidden_dim,
+                num_attention_heads=num_attention_heads,
+                layer_number=i + 1,
+                fuse_qkv_params=False,
+            )
+            for i in range(processor_layers)
+        ]
+        self.processor_layers = nn.ModuleList(layers)
+
+    def forward(
+        self,
+        nfeat: Tensor,
+    ) -> Tensor:
+        nfeat = nfeat.unsqueeze(1)
+        # TODO make sure reshaping the last dim to (h, d) is done automatically in the transformer layer
+        for module in self.processor_layers:
+            nfeat = module(
+                nfeat,
+                attention_mask=self.mask,
+                self_attn_mask_type="arbitrary",
+            )
+
+        return torch.squeeze(nfeat, 1)