+++ test/models/graphcast/test_cugraphops.py
@@ -52,7 +52,7 @@ def test_cugraphops(
     np.random.seed(0)
 
     model = GraphCastNet(
-        multimesh_level=1,
+        mesh_level=1,
         input_res=(res_h, res_w),
         input_dim_grid_nodes=num_channels,
         input_dim_mesh_nodes=3,
@@ -71,7 +71,7 @@ def test_cugraphops(
     fix_random_seeds()
 
     model_dgl = GraphCastNet(
-        multimesh_level=1,
+        mesh_level=1,
         input_res=(res_h, res_w),
         input_dim_grid_nodes=num_channels,
         input_dim_mesh_nodes=3,